% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getgeodataApp}
\alias{getgeodataApp}
\title{function for download geodata}
\usage{
getgeodataApp(geopath = base::readline("path to dep6:"))
}
\arguments{
\item{geopath}{path to d6 geodata server, can be set before}
}
\value{
a shiny app for downloading geodata from our d6 server
}
\description{
Use this function to browse and download geodata from our d6 server. Use this function within you project, because the data can only be saved within you directory.
}
\examples{
\dontrun{
getgeodataApp()
}
}
