% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getgeodataApp}
\alias{getgeodataApp}
\title{function for download geodata}
\usage{
getgeodataApp(geopath = base::readline("path to dep6:"))
}
\arguments{
\item{geopath}{path to d6 geodata server}
}
\value{
a shiny app for downloading geodata from our d6 server
}
\description{
function for download geodata
}
\examples{
\dontrun{
getgeodataApp()
}
}
